Simplifies connection tests and removes unused bits.
Adds OnAcked and OnLost handlers to Reliability type.
These can be assigned a function pointer to handle the acknowledgement
and loss of individual packets. Our library does not resend lost
packets, because this is exactly what makes TCP slow. We leave the
way this is handled to the host application.