Adds the ident sub package. It adds a peer identification layer on top of a basic xudp.Connection.
For this purpose, it includes a 32 byte SHA256 hash of a peer's
internal IP/port into each packet. The receiving end combines this
with the peer's public IP in an Endpoint struct. The Endpoint.String()
method returns a 32 byte SHA256 hash of the ID and the public IP.
It can be used to uniquely identify every peer, even if one or more
of them come from behind a NAT router/firewall.