Completely revamps the package structure to be completely and reliably modular.
xudp.Connection is no more than a wrapper around basic
Go UDP functionality. Any features like protocolID, peer identification
or reliability can be added through plugins which must be
registered with the connection.

This allows the host to cherry-pick only those components they
require for a given purpose. The packet construction and handling
is all completely transparent. This approach also makes it easy
for the host to implement their own plugins which perform
whatever actions are required on incoming and outgoing packet data.